refactor: Clean code[843652]

+++ qmmp/src/plugins/Effect/muffler/mufflerplugin.cpp
@@ -19,7 +19,7 @@ void MufflerPlugin::applyEffect(Buffer *b)
     }
 
     float *data = b->data;
-    for(uint i = 0; i < b->samples; i += 2)
+    for(size_t i = 0; i < b->samples; i += 2)
     {
         data[i] -= data[i + 1];
         data[i + 1] = data[i];
